add torch save/load

+++ models_torch.py
@@ -467,10 +467,10 @@ class Agent:
 
     def train_minibatch(self, batch_size=16):
         state, action, reward, next_state = self.sample_minibatch(batch_size)
-        x = {'state': torch.from_numpy(np.concatenate(state)),
-            'action': torch.from_numpy(np.concatenate(action)), 
-            'reward': torch.from_numpy(np.array(reward)).view(-1, 1),
-            'next_state': torch.from_numpy(np.concatenate(next_state))}
+        x = {'state': torch.from_numpy(np.concatenate(state)).double(),
+            'action': torch.from_numpy(np.concatenate(action)).double(), 
+            'reward': torch.from_numpy(np.array(reward)).view(-1, 1).double(),
+            'next_state': torch.from_numpy(np.concatenate(next_state)).double()}
         self.model.train_step(x)
         return self.model.critic_loss(x)
 
@@ -518,15 +518,14 @@ class Agent:
                 self.history['batch_loss'].append(batch_loss)
             self.history['rewards'].append(np.sum(episode_rewards))
             self.history['dqn_loss'].append(self.get_q_loss())
-        #self.model.save_target_model('test_model')
         print('Finishing Training: {:.2f}s'.format(time.time() - training_start))
 
     def get_q_loss(self):
         state, action, reward, next_state = [[sample[i] for sample in self.experience_buffer] for i in range(4)]
-        return self.model.critic_loss({'state': torch.from_numpy(np.concatenate(state)),
-                                    'action': torch.from_numpy(np.concatenate(action)), 
-                                    'reward': torch.from_numpy(np.array(reward)).view(-1, 1),
-                                    'next_state': torch.from_numpy(np.concatenate(next_state))})
+        return self.model.critic_loss({'state': torch.from_numpy(np.concatenate(state)),double(),
+                                    'action': torch.from_numpy(np.concatenate(action)),double(), 
+                                    'reward': torch.from_numpy(np.array(reward)).view(-1, 1),double(),
+                                    'next_state': torch.from_numpy(np.concatenate(next_state)),double()})
     def save_target_model(self):
         self.model.save_target_model()
